+++ src/database/db_schema.rs
@@ -25,7 +25,6 @@ pub(crate) fn create_price() -> String {
                     .auto_increment()
                     .primary_key(),
             )
-            .col(ColumnDef::new(PriceIden::SecurityId).integer())
             .col(ColumnDef::new(PriceIden::Date).string())
             .col(ColumnDef::new(PriceIden::Time).string().null())
             .col(ColumnDef::new(PriceIden::Value).integer())